fix: renamed the file mixpanelDef to mixpanelEvents and moved everything into a new folder called mixpanel
