+++ src/routes/__tests__/App.test.tsx
@@ -43,7 +43,7 @@ import userEvent from "@testing-library/user-event";
 import fetchMock from "jest-fetch-mock";
 import { renderWithReduxProvider } from "../../utils/testing/testRenderProviders";
 import * as helper from "../../utils/api/helper";
-import * as mixpanel from "../../utils/config/mixpanelHelperInit";
+import * as mixpanel from "../../utils/mixpanel/mixpanelHelperInit";
 import { App } from "../../App";
 
 // Mock MUI Italia theme
