+++ src/utils/__tests__/transactions/transactionHelper.test.ts
@@ -1,4 +1,4 @@
-jest.mock("../../config/mixpanelHelperInit", () => ({
+jest.mock("../../mixpanel/mixpanelHelperInit", () => ({
   mixpanel: { track: jest.fn() },
 }));
